+++ src/data/connections.ts
@@ -1,0 +1,15 @@
+import knex from 'knex';
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+export const connection = knex({
+	client: "mysql",
+	connection: {
+    host: process.env.DB_HOST,
+    port: 3306,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASS,
+    database: process.env.DB_NAME
+  }
+});
