+++ src/index.ts
@@ -1,17 +1,4 @@
-import express, {Express} from 'express';
-import cors from 'cors';
-import {AddressInfo} from 'net';
+import { criarDocente } from "./endpoints/criarDocente";
+import app from "./app";
 
-const app: Express = express();
-
-app.use(cors());
-app.use(express.json());
-
-const server = app.listen(process.env.PORT || 3000, () => {
-    if (server) {
-        const address = server.address() as AddressInfo;
-        console.log(`Server is running in http://localhost:${address.port}`);
-    } else {
-        console.error(`Failure upon starting server.`);
-    }
-});
+app.post('/create/teacher', criarDocente);
